+++ src/index.css
@@ -1,5 +1,6 @@
 @import url('https://fonts.googleapis.com/css2?family=Manrope:wght@200..800&family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&family=Roboto+Slab:wght@100..900&display=swap');
 @import "tailwindcss";
+@import url('https://fonts.googleapis.com/css2?family=Libertinus+Serif:ital,wght@0,400;0,600;0,700;1,400;1,600;1,700&display=swap');
 
 ::selection {
   background-color: #f0e68c; /* Light yellow */
@@ -52,4 +53,10 @@
   background-color: #b07a3c;
   color: #fff;
   transform: translateY(-50%) scale(1.1);
+}
+
+.libertinus-serif-bold {
+  font-family: 'Libertinus Serif', serif;
+  font-weight: 700;
+  font-style: normal;
 }
